% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titanic.complete.R
\docType{package}
\name{titanic.complete}
\alias{titanic.complete}
\alias{titanic.complete-package}
\title{titanic.complete}
\description{
This package attempts to provide a complete dataset of all people on the Titanic,
including both passengers and crew. The data are provided in the data frame
\code{\link{titanic_complete}}.
}
\seealso{
\code{\link{titanic_complete}}
}
