% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{titanic_complete}
\alias{titanic_complete}
\title{Complete dataset of titanic passengers and crew.}
\format{A data frame with 2179 rows and 14 variables:
\describe{
\item{\code{Name}}{Full name of the person}
\item{\code{Age}}{Age (in years) of the person}
\item{\code{...}}{to be completed}
}}
\usage{
titanic_complete
}
\description{
Detailed description of the dataset and its origin goes here.
}
\keyword{datasets}
